requirements.txt
pandas
requests
matplotlib
python-dotenv==1.0.0
PyGithub==2.1.1
beautifulsoup4==4.12.2